lambda/producer/requirements.txt
kafka-python==2.0.2
aws-msk-iam-sasl-signer-python